legal-callbot/brain/requirements.txt
# ── LLM ──
google-genai>=1.0.0

# ── Embedding (BGE-M3) ──
FlagEmbedding>=1.2.0
transformers>=4.44.0,<4.46.0
torch>=2.0.0

# ── Vector DB ──
qdrant-client>=1.7.0

# ── API Server ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# ── Utilities ──
python-dotenv>=1.0.0
